requirements/base.txt
channels==3.0.5
daphne==3.0.2
dj-rest-auth==2.2.5
Django==3.2.16
django-allauth==0.51.0
djangorestframework==3.14.0
